% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/connect_source.R
\name{connect_attribute}
\alias{connect_attribute}
\title{Make a git connection to the attribute}
\usage{
connect_attribute(result.channel, username, password)
}
\arguments{
\item{result.channel}{An open RODBC connection to the results database}

\item{username}{The optional username used in case of  https authentication. Ignored when \code{key} is provided.}

\item{password}{The password required for the ssh key or the username. Should be missing when the ssh-key doesn't require a password.}
}
\description{
Make a git connection to the attribute
}

