% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_source.R
\name{connect_flemish_source}
\alias{connect_flemish_source}
\title{Make an ODBC connection to the Flemish source data}
\usage{
connect_flemish_source(result.channel)
}
\arguments{
\item{result.channel}{An open RODBC connection to the results database}
}
\description{
Make an ODBC connection to the Flemish source data
}
