% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_source.R
\name{connect_nbn}
\alias{connect_nbn}
\title{Make an ODBC connection to the NBN database}
\usage{
connect_nbn(result.channel)
}
\arguments{
\item{result.channel}{An open RODBC connection to the results database}
}
\description{
Make an ODBC connection to the NBN database
}
