% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_source.R
\name{connect_raw}
\alias{connect_raw}
\title{Make a git connection to the raw data}
\usage{
connect_raw(result.channel, username, password,
  commit.user = "watervogelanalysis", commit.email = "bmk@inbo.be")
}
\arguments{
\item{result.channel}{An open RODBC connection to the results database}

\item{username}{The optional username used in case of  https authentication.
Ignored when \code{key} is provided.}

\item{password}{The password required for the ssh key or the username. Should
be missing when the ssh-key doesn't require a password.}

\item{commit.user}{the name of the user how will commit}

\item{commit.email}{the email of the user how will commit}
}
\description{
Make a git connection to the raw data
}

