% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/connect_watervogel.R
\name{connect_watervogel}
\alias{connect_watervogel}
\title{Opens an ODBC connection to the 'Watervogel' database}
\usage{
connect_watervogel()
}
\description{
Opens an ODBC connection to the 'Watervogel' database
}

