% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource_id.R
\name{datasource_id_raw}
\alias{datasource_id_raw}
\title{Get the id of the data source for the raw data}
\usage{
datasource_id_raw(result_channel)
}
\arguments{
\item{result_channel}{An open \code{RODBC} connection to the results database}
}
\description{
The raw data is the combination of the Flemish and Walloon source data
}
