% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource_id.R
\name{datasource_id_result}
\alias{datasource_id_result}
\title{Get the id of the data source for the raw data}
\usage{
datasource_id_result(result_channel, develop = FALSE)
}
\arguments{
\item{result_channel}{An open \code{RODBC} connection to the results database}

\item{develop}{logical.
Indicate whether the use the development database TRUE or the production
database.}
}
\description{
The raw data is the combination of the Flemish and Walloon source data
}
