% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_result.R
\name{extract_results}
\alias{extract_results}
\title{Extract results from a model}
\usage{
extract_results(x, ...)
}
\arguments{
\item{x}{The model to extract results from}

\item{...}{Additional arguments passed to the extraction method}
}
\description{
Extract results from a model
}
