% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gbif.R
\name{get_gbif}
\alias{get_gbif}
\title{Get GBIF data for a species}
\usage{
get_gbif(scientific, language = c("nld", "eng", "fra", "deu"), verbose = TRUE)
}
\arguments{
\item{scientific}{The scientific name of the species}

\item{verbose}{If TRUE, display progress messages}
}
\value{
A data frame with the vernacular names in different languages
}
\description{
Get GBIF data for a species
}
