% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/import_walloon_raw_data.R
\name{import_walloon_raw_data}
\alias{import_walloon_raw_data}
\title{Add the raw data from Wallonia to the git repository}
\usage{
import_walloon_raw_data(location.file, visit.file, data.file, path = ".",
  username, password)
}
\arguments{
\item{location.file}{file with details on the location}

\item{visit.file}{file with details on the visits to each location}

\item{data.file}{file with observed species at each visit}

\item{path}{directory were the above files are stored}

\item{username}{the username for the git repository}

\item{password}{the user's password for the git repository}
}
\description{
This functions reads the files and performs some basic checks on them. See the details section for the required format of the files.
}
\details{
All files must textfile with ';' separated fields and '.' to indicate decimal points

\code{location.file} must have the following fields:
\itemize{
  \item{\code{CODE_ULT} The id of the location}
  \item{\code{Nom_site} The name of the location}
  \item{\code{ZPS} 1 if the location is a Special Procection Area, otherwise 0}
  \item{\code{Area} The surface area of the location in square m}
  \item{\code{Xlamb} The X coordinate of the location, Belgium Lambert 72 system}
  \item{\code{Ylamb} The Y coordinate of the location, Belgium Lambert 72 system}
}

\code{visit.file} must have the following fields:
\itemize{
  \item{\code{ID_visit} The is of the visit to the location}
  \item{\code{CODE_ULT} The id of the location}
  \item{\code{DATE} The date of the visit to the location}
}

\code{visit.file} must have the following fields:
\itemize{
  \item{\code{ID_visit} The is of the visit to the location}
  \item{\code{TAXPRIO} The scientific name of the species. Only exactly matches from \code{\link{read_specieslist}} with \code{limit = FALSE} are retained. Non-matching values are displayed in a warning.}
  \item{\code{SommeDeN} The observed number of animals}
}
}

