% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_analysis_aggregate_wintermax.R
\name{prepare_analysis_agg_max}
\alias{prepare_analysis_agg_max}
\title{Create aggregation objects for imputed counts}
\usage{
prepare_analysis_agg_max(aggregation, analysis_path, verbose = TRUE)
}
\arguments{
\item{aggregation}{an \code{n2kAggregate} object}

\item{analysis_path}{Path to store the analysis files.
Must be either an existing local file path or an object of the \code{s3_bucket}
class.}

\item{verbose}{Display a progress bar when TRUE (default)}
}
\description{
Create aggregation objects for imputed counts
}
