% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_analysis_aggregate.R
\name{prepare_analysis_aggregate}
\alias{prepare_analysis_aggregate}
\title{Create aggregation objects for imputed counts}
\usage{
prepare_analysis_aggregate(analysis.path, imputations, verbose = TRUE)
}
\arguments{
\item{analysis.path}{Path to store the analysis files. Must be either an existing local file path or an object of the \code{s3_bucket} class.}

\item{imputations}{a data.frame with the imputations per location group}

\item{verbose}{Display a progress bar when TRUE (default)}
}
\description{
Create aggregation objects for imputed counts
}
