% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_analysis_aggregate_ni.R
\name{prepare_analysis_aggregate_ni}
\alias{prepare_analysis_aggregate_ni}
\title{Create aggregation objects for imputed counts while ignoring the missing data}
\usage{
prepare_analysis_aggregate_ni(count, analysis_path, raw_repo, verbose = TRUE)
}
\arguments{
\item{count}{An \code{n2kInla} object holding the count data.}

\item{analysis_path}{Path to store the analysis files.
Must be either an existing local file path or an object of the \code{s3_bucket}
class.}

\item{raw_repo}{a git_repository object to write the output to}

\item{verbose}{Display a progress bar when TRUE (default)}
}
\description{
Create aggregation objects for imputed counts while ignoring the missing data
}
