% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_analysis_composite.R
\name{prepare_analysis_composite}
\alias{prepare_analysis_composite}
\title{Prepare the analysis files for the composite indices}
\usage{
prepare_analysis_composite(
  species_group_id,
  location_group_id,
  models,
  base,
  project,
  scheme_id,
  seed = 20070315,
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{models}{a dataframe with the parent models.}

\item{base}{The base location to store the model.}

\item{project}{Will be a relative path within the base location.}

\item{overwrite}{Should an existing object be overwritten?
Defaults to \code{TRUE}.}
}
\description{
Prepare the analysis files for the composite indices
}
