% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_analysis_imputation.R
\name{prepare_analysis_imputation}
\alias{prepare_analysis_imputation}
\title{Create the analysis dataset based on the available raw data}
\usage{
prepare_analysis_imputation(speciesgroupspecies, location, analysis_path,
  raw_repo, seed = 19790402, verbose = TRUE)
}
\arguments{
\item{speciesgroupspecies}{a data.frame with the current species group and the associates species}

\item{location}{a data.frame with the locations and location groups}

\item{analysis_path}{Path to store the analysis files. Must be either an existing local file path or an object of the \code{s3_bucket} class.}

\item{raw_repo}{a git_repository object to write the output to}

\item{seed}{the seed for the random number generator. Defaults to 19790402 which refers to "Council Directive 79/409/EEC of 2 April 1979 on the conservation of wild birds".}

\item{verbose}{Display a progress bar when TRUE (default)}
}
\value{
A data.frame with the species id number of rows in the analysis dataset, number of precenses in the analysis datset and SHA-1 of the analysis dataset or NULL if not enough data.
}
\description{
The analysis dataset is saved to a rda file with the SHA-1 as name.
}
