% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_analysis_index.R
\name{prepare_analysis_index}
\alias{prepare_analysis_index}
\title{Generate the models to apply on the aggregated sets}
\usage{
prepare_analysis_index(
  aggregation,
  analysis_path,
  month = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{aggregation}{an \code{n2kAggregate} object}

\item{analysis_path}{Path to store the analysis files.
Must be either an existing local file path or an object of the \code{s3_bucket}
class.}

\item{month}{Model the month effect.
Defaults to \code{TRUE}.}

\item{verbose}{Display a progress bar when TRUE (default)}
}
\description{
Generate the models to apply on the aggregated sets
}
