% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_analysis_model.R
\name{prepare_analysis_model}
\alias{prepare_analysis_model}
\title{Generate the models to apply on the aggregated sets}
\usage{
prepare_analysis_model(aggregation, analysis.path, verbose = TRUE)
}
\arguments{
\item{aggregation}{the output of \code{prepare_analysis_aggregate}}

\item{analysis.path}{Path to store the analysis files. Must be either an existing local file path or an object of the \code{s3_bucket} class.}

\item{verbose}{Display a progress bar when TRUE (default)}
}
\description{
Generate the models to apply on the aggregated sets
}
