% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_dataset.R
\name{prepare_dataset}
\alias{prepare_dataset}
\title{Prepare the raw datasets and save them to the git repository}
\usage{
prepare_dataset(
  raw_repo,
  walloon_repo,
  flemish_channel,
  verbose = TRUE,
  first_year = 1992,
  latest_year = as.integer(format(Sys.time(), "\%Y")),
  strict = TRUE
)
}
\arguments{
\item{raw_repo}{a git_repository object to write the output to}

\item{walloon_repo}{a git_repository object to the Walloon source data}

\item{flemish_channel}{a DBI connection to the Flemish database}

\item{verbose}{Display a progress bar when TRUE (default)}

\item{first_year}{first winter to import. defaults to 1992}

\item{latest_year}{latest winter to import.
Winter 2019 is defined as 2018-10-01 until 2019-03-31.
Defaults the winter prior to last firth of July.
2019-06-30 becomes 2018, 2019-07-01 becomes 2019.}

\item{strict}{What to do when the metadata changes. \code{strict = FALSE}
overwrites the data and the metadata with a warning listing the changes,
\code{strict = TRUE} returns an error and leaves the data and metadata as is.
Defaults to \code{TRUE}.}
}
\description{
The raw data is written to the git repository.
All changes are always staged and committed.
The commit is pushed when both username and password are provided.
}
\examples{
\dontrun{
 prepare_dataset()
}
}
