% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_dataset_location.R
\name{prepare_dataset_location}
\alias{prepare_dataset_location}
\title{Read the locations for the raw data source, save them to the git repository
and the results database}
\usage{
prepare_dataset_location(
  flemish_channel,
  walloon_repo,
  raw_repo,
  scheme_id = "watervogels",
  first_date,
  latest_date = as.POSIXct(Sys.time())
)
}
\arguments{
\item{flemish_channel}{a DBI connection to the Flemish database}

\item{walloon_repo}{a git_repository object to the Walloon source data}

\item{raw_repo}{a git_repository object to write the output to}

\item{scheme_id}{the id of the scheme}

\item{first_date}{earliest records to take into account}

\item{latest_date}{until which date should we import the species list}
}
\value{
a data.frame with the locations
}
\description{
Read the locations for the raw data source, save them to the git repository
and the results database
}
