% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_dataset_observation.R
\name{prepare_dataset_observation}
\alias{prepare_dataset_observation}
\title{Read the observations for the raw data source, save them to the git
repository}
\usage{
prepare_dataset_observation(
  this_species,
  location,
  flemish_channel,
  walloon_repo,
  raw_repo,
  latest_year,
  verbose = TRUE
)
}
\arguments{
\item{this_species}{the species in this species group}

\item{location}{a data frame with the full list of locations}

\item{flemish_channel}{a DBI connection to the Flemish database}

\item{walloon_repo}{a git_repository object to the Walloon source data}

\item{raw_repo}{a git_repository object to write the output to}

\item{latest_year}{latest winter to import.
Winter 2019 is defined as 2018-10-01 until 2019-03-31.
Defaults the winter prior to last firth of July.
2019-06-30 becomes 2018, 2019-07-01 becomes 2019.}

\item{verbose}{Display a progress bar when TRUE (default)}
}
\description{
Read the observations for the raw data source, save them to the git
repository
}
