% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_dataset_species.R
\name{prepare_dataset_species}
\alias{prepare_dataset_species}
\title{Read the species for the raw data source, save them to the results database}
\usage{
prepare_dataset_species(
  raw_repo,
  flemish_channel,
  walloon_repo,
  first_date,
  latest_date
)
}
\arguments{
\item{raw_repo}{a git_repository object to write the output to}

\item{flemish_channel}{a DBI connection to the Flemish database}

\item{walloon_repo}{a git_repository object to the Walloon source data}

\item{first_date}{earliest records to take into account}

\item{latest_date}{until which date should we import the species list}
}
\value{
the species group constraint information
}
\description{
Read the species for the raw data source, save them to the results database
}
