% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw2result.R
\name{raw2result}
\alias{raw2result}
\title{Transform raw species and locationgroup data to result data}
\usage{
raw2result(raw_data, root)
}
\arguments{
\item{raw_data}{Character string with path to the raw data directory.}

\item{root}{Character string with path to the results directory.}
}
\description{
Transform raw species and locationgroup data to result data
}
