% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_location.R
\name{read_location}
\alias{read_location}
\title{read the dataset of locations from the database}
\usage{
read_location(flemish_channel, walloon_repo, first_date, latest_date)
}
\arguments{
\item{flemish_channel}{a DBI connection to the Flemish database}

\item{walloon_repo}{a git_repository object to the Walloon source data}

\item{first_date}{earliest records to take into account}

\item{latest_date}{until which date should we import the species list}
}
\description{
read the dataset of locations from the database
}
