% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_observation.R
\name{read_observation}
\alias{read_observation}
\title{Read the Flemish observations from a species}
\usage{
read_observation(species_id, latest_year, flemish_channel)
}
\arguments{
\item{species_id}{The id of the species}

\item{latest_year}{latest winter to import. Winter 2019 is defined as 2018-10-01 until 2019-03-31}

\item{flemish_channel}{a DBI connection to the Flemish database}
}
\value{
A \code{data.frame} with observations. \code{Complete= 1} indicates that the entire location was surveyed.
}
\description{
The selection uses several constraints. The user defined constraints are \code{first.winter} and \code{species.covered}. Following contraints are imposed at the same time as the user defined contraints:
\itemize{
\item The observation is validated
\item The observation is a midmonthly observation
\item The observation is not 'unobserved'
\item The location is 'active' and the observation date is within the 'active' period of the location
\item The observation is between october and march
}
}
