% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_specieslist.R
\name{read_specieslist}
\alias{read_specieslist}
\title{Read the relevant species list}
\usage{
read_specieslist(
  raw_repo,
  flemish_channel,
  walloon_repo,
  first_date,
  latest_date
)
}
\arguments{
\item{raw_repo}{a git_repository object to write the output to}

\item{flemish_channel}{a DBI connection to the Flemish database}

\item{walloon_repo}{a git_repository object to the Walloon source data}

\item{first_date}{earliest records to take into account}

\item{latest_date}{until which date should we import the species list}
}
\description{
Read the relevant species list
}
