% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_duplicate_observation.R
\name{remove_duplicate_observation}
\alias{remove_duplicate_observation}
\title{Remove duplicate observations}
\usage{
remove_duplicate_observation(observation)
}
\arguments{
\item{observation}{a data.frame with observations}
}
\value{
a list with components \code{Observation} and \code{Duplicate}. \code{Observation} holds the cleaned observations. \code{Duplicate} is a data.frame with the ObservationID and DatasourceID of duplicated records.
}
\description{
Duplicate observations are combinations of LocationID, Year and fMonth for which multiple rows exist.
}

