% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scheme_id.R
\name{scheme_id}
\alias{scheme_id}
\title{Get the id of the scheme}
\usage{
scheme_id(result.channel)
}
\arguments{
\item{result.channel}{An open RODBC connection to the results database}
}
\description{
Get the id of the scheme
}

