% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/select_relevant.R
\name{select_relevant}
\alias{select_relevant}
\title{Select the relevant observation}
\usage{
select_relevant(observation)
}
\arguments{
\item{observation}{the \code{data.frame} with observations}
}
\description{
Relevant observations have
\itemize{
 \item At least 4 presences per location
 \item Only months who's average is at least 5% of the month with the highst abundance
 \item Only locations where the species was present during at least 3 winters
 \item No winters without any presences at the beginning or end of the dataset
}
}

