% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_relevant_import.R
\name{select_relevant_import}
\alias{select_relevant_import}
\title{Select the relevant observations when importing the raw data}
\usage{
select_relevant_import(observation)
}
\arguments{
\item{observation}{the \code{data.frame} with observations}
}
\description{
Relevant observations have
\itemize{
 \item At least 4 presences per location
 \item Only locations where the species was present during at least 3 winters
 \item No winters without any presences at the beginning or end of the dataset
}
}
