% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/connect_source.R
\name{connect_attribute}
\alias{connect_attribute}
\title{Make a git connection to the attribute}
\usage{
connect_attribute(result.channel, username, password)
}
\arguments{
\item{result.channel}{An open RODBC connection to the results database}

\item{username}{the username in case the ConnectMethod is "Credentials supplied by the user running the report". Ignored in all other cases.}

\item{password}{the password to be used in combination with the username.}
}
\description{
Make a git connection to the attribute
}

