% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource_id.R
\name{datasource_id_flanders}
\alias{datasource_id_flanders}
\title{Get the id of the data source for the Flemish source data}
\usage{
datasource_id_flanders(result.channel)
}
\arguments{
\item{result.channel}{An open RODBC connection to the results database}
}
\description{
Get the id of the data source for the Flemish source data
}
