% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasource_id.R
\name{datasource_id_raw}
\alias{datasource_id_raw}
\title{Get the id of the data source for the raw data}
\usage{
datasource_id_raw(result.channel)
}
\arguments{
\item{result.channel}{An open RODBC connection to the results database}
}
\description{
The raw data is the combination of the Flemish and Walloon source data
}

