% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource_id.R
\name{datasource_id_wallonia}
\alias{datasource_id_wallonia}
\title{Get the id of the data source for the Walloon source data}
\usage{
datasource_id_wallonia(result_channel)
}
\arguments{
\item{result_channel}{An open RODBC connection to the results database}
}
\description{
Get the id of the data source for the Walloon source data
}
