% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_walloon_source_data.R
\name{import_walloon_source_data}
\alias{import_walloon_source_data}
\title{Add the raw data from Wallonia to the git repository}
\usage{
import_walloon_source_data(
  location_file,
  visit_file,
  species_file,
  data_file,
  path = ".",
  walloon_repo,
  strict = TRUE
)
}
\arguments{
\item{location_file}{file with details on the location}

\item{visit_file}{file with details on the visits to each location}

\item{species_file}{file with all species}

\item{data_file}{file with observed species at each visit}

\item{path}{directory were the above files are stored}

\item{walloon_repo}{a git_repository object to the Walloon source data}

\item{strict}{What to do when the metadata changes. \code{strict = FALSE}
overwrites the data and the metadata with a warning listing the changes,
\code{strict = TRUE} returns an error and leaves the data and metadata as is.
Defaults to \code{TRUE}.}
}
\description{
This functions reads the files and performs some basic checks on them.
See the details section for the required format of the files.
The median date is used in case of multiple dates per visit id.
The maximum is used in case of multiple observations per visit id.
}
