% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prepare_analysis.R
\name{prepare_analysis}
\alias{prepare_analysis}
\title{Prepare all datasets and a to do list of models}
\usage{
prepare_analysis(output.path = ".", raw.connection)
}
\arguments{
\item{output.path}{Path to store the analysis files}

\item{raw.connection}{a git-connection object to write the output to}
}
\description{
Prepare all datasets and a to do list of models
}

