% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prepare_analysis.R
\name{prepare_analysis}
\alias{prepare_analysis}
\title{Prepare all datasets and a to do list of models}
\usage{
prepare_analysis(analysis.path = ".", raw.connection, scheme.id)
}
\arguments{
\item{analysis.path}{Path to store the analysis files}

\item{raw.connection}{a git-connection object to write the output to}

\item{scheme.id}{the id of the scheme}
}
\description{
Prepare all datasets and a to do list of models
}

