% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_analysis_model.R
\name{prepare_analysis_model}
\alias{prepare_analysis_model}
\title{Generate the models to apply on the aggregated sets}
\usage{
prepare_analysis_model(aggregation, analysis.path, verbose = TRUE)
}
\description{
Generate the models to apply on the aggregated sets
}
