% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prepare_dataset.R
\name{prepare_dataset}
\alias{prepare_dataset}
\title{Prepare the raw datasets and save them to the git repository}
\usage{
prepare_dataset(verbose = TRUE)
}
\arguments{
\item{verbose}{Display a progress bar when TRUE (default)}
}
\description{
Prepare the raw datasets and save them to the git repository
}
\examples{
\dontrun{
 prepare_dataset()
}
}

