% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prepare_dataset.R
\name{prepare_dataset}
\alias{prepare_dataset}
\title{Prepare the raw datasets and save them to the git repository}
\usage{
prepare_dataset(username, password, scheme.id = get_scheme_id("Watervogels"),
  verbose = TRUE, develop = TRUE)
}
\arguments{
\item{username}{the username for the git repository}

\item{password}{the user's password for the git repository}

\item{scheme.id}{the id of the scheme}

\item{verbose}{Display a progress bar when TRUE (default)}

\item{develop}{Logical value. Indicates the location ton the results database}
}
\description{
The raw data is written to the git repository. All changes are always staged and committed. The commit is pushed when both username and password are provided.
}
\examples{
\dontrun{
 prepare_dataset()
}
}

