% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_dataset.R
\name{prepare_dataset}
\alias{prepare_dataset}
\title{Prepare the raw datasets and save them to the git repository}
\usage{
prepare_dataset(scheme.id, raw.connection, result.channel, attribute.connection,
  walloon.connection, flemish.channel, nbn.channel, verbose = TRUE)
}
\arguments{
\item{scheme.id}{the id of the scheme}

\item{raw.connection}{a git-connection object to write the output to}

\item{result.channel}{An open RODBC connection to the results database}

\item{attribute.connection}{a git-connection object to the attributes}

\item{walloon.connection}{a git-connection object to the Walloon source data}

\item{flemish.channel}{An open ODBC connection to the source database}

\item{nbn.channel}{an RODBC connection to the NBN database}

\item{verbose}{Display a progress bar when TRUE (default)}
}
\description{
The raw data is written to the git repository. All changes are always staged and committed. The commit is pushed when both username and password are provided.
}
\examples{
\dontrun{
 prepare_dataset()
}
}
