% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prepare_dataset_location.R
\name{prepare_dataset_location}
\alias{prepare_dataset_location}
\title{Read the locations for the raw datasource, save them to the git repository and the results database}
\usage{
prepare_dataset_location(scheme.id, result.channel, flemish.channel,
  walloon.connection, raw.connection)
}
\arguments{
\item{scheme.id}{the id of the scheme}

\item{result.channel}{An open RODBC connection to the results database}

\item{flemish.channel}{An open ODBC connection to the source database}

\item{walloon.connection}{a git-connection object to the Walloon source data}

\item{raw.connection}{a git-connection object to write the output to}
}
\value{
a data.frame with the locations
}
\description{
Read the locations for the raw datasource, save them to the git repository and the results database
}

