% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prepare_dataset_location.R
\name{prepare_dataset_location}
\alias{prepare_dataset_location}
\title{Read the locations for the raw datasource, save them to the git repository and the results database}
\usage{
prepare_dataset_location(scheme.id = odbc_get_id(table = "Scheme", variable =
  "Description", value = "Watervogels", develop = develop), develop = develop)
}
\arguments{
\item{scheme.id}{the id of the scheme}

\item{develop}{Logical value. Indicates the location ton the results database}
}
\value{
a data.frame with the locations
}
\description{
Read the locations for the raw datasource, save them to the git repository and the results database
}

