% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prepare_dataset_observation.R
\name{prepare_dataset_observation}
\alias{prepare_dataset_observation}
\title{Read the observations for the raw datasource, save them to the git repository and the results database}
\usage{
prepare_dataset_observation(this.constraint, location,
  scheme.id = odbc_get_id(table = "Scheme", variable = "Description", value =
  "Watervogels", develop = develop), develop = TRUE)
}
\arguments{
\item{this.constraint}{the constraints for this species group}

\item{location}{a data frame with the full list of locations}

\item{scheme.id}{the id of the scheme}

\item{develop}{Logical value. Indicates the location ton the results database}
}
\description{
Read the observations for the raw datasource, save them to the git repository and the results database
}

