% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_dataset_observation.R
\name{prepare_dataset_observation}
\alias{prepare_dataset_observation}
\title{Read the observations for the raw datasource, save them to the git repository and the results database}
\usage{
prepare_dataset_observation(
  this_species,
  location,
  location_group_id,
  flemish_channel,
  walloon_repo,
  result_channel,
  raw_repo,
  dataset
)
}
\arguments{
\item{this_species}{the species in this species group}

\item{location}{a data frame with the full list of locations}

\item{location_group_id}{the ID of the location group}

\item{flemish_channel}{a DBI connection to the Flemish database}

\item{walloon_repo}{a git_repository object to the Walloon source data}

\item{result_channel}{An open RODBC connection to the results database}

\item{raw_repo}{a git_repository object to write the output to}

\item{dataset}{a data frame with a}
}
\description{
Read the observations for the raw datasource, save them to the git repository and the results database
}
