% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_dataset_observation.R
\name{prepare_dataset_observation}
\alias{prepare_dataset_observation}
\title{Read the observations for the raw datasource, save them to the git repository and the results database}
\usage{
prepare_dataset_observation(this.constraint, location, location_group_id,
  flemish.channel, walloon.connection, result.channel, raw.connection, dataset)
}
\arguments{
\item{this.constraint}{the constraints for this species group}

\item{location}{a data frame with the full list of locations}

\item{location_group_id}{the ID of the location group}

\item{flemish.channel}{An open ODBC connection to the source database}

\item{walloon.connection}{a git-connection object to the Walloon source data}

\item{result.channel}{An open RODBC connection to the results database}

\item{raw.connection}{a git-connection object to write the output to}

\item{dataset}{a data frame with a}
}
\description{
Read the observations for the raw datasource, save them to the git repository and the results database
}
