% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prepare_dataset_species.R
\name{prepare_dataset_species}
\alias{prepare_dataset_species}
\title{Read the species for the raw datasource, save them to the results database}
\usage{
prepare_dataset_species(scheme.id = odbc_get_id(table = "Scheme", variable =
  "Description", value = "Watervogels", develop = develop), develop = develop)
}
\arguments{
\item{scheme.id}{the id of the scheme}

\item{develop}{Logical value. Indicates the location ton the results database}
}
\value{
the species group constraint information
}
\description{
Read the species for the raw datasource, save them to the results database
}

