% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_dataset_species.R
\name{prepare_dataset_species}
\alias{prepare_dataset_species}
\title{Read the species for the raw datasource, save them to the results database}
\usage{
prepare_dataset_species(raw.connection, result.channel, flemish.channel,
  attribute.connection, walloon.connection, nbn.channel, scheme.id)
}
\arguments{
\item{raw.connection}{a git-connection object to write the output to}

\item{result.channel}{An open RODBC connection to the results database}

\item{flemish.channel}{An open ODBC connection to the source database}

\item{attribute.connection}{a git-connection object to the attributes}

\item{walloon.connection}{a git-connection object to the Walloon source data}

\item{nbn.channel}{an RODBC connection to the NBN database}

\item{scheme.id}{the id of the scheme}
}
\value{
the species group constraint information
}
\description{
Read the species for the raw datasource, save them to the results database
}
