% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prepare_dataset_species.R
\name{prepare_dataset_species}
\alias{prepare_dataset_species}
\title{Read the species for the raw datasource, save them to the results database}
\usage{
prepare_dataset_species(scheme.id, result.channel, flemish.channel,
  attribute.connection, walloon.connection)
}
\arguments{
\item{scheme.id}{the id of the scheme}

\item{result.channel}{An open RODBC connection to the results database}

\item{flemish.channel}{An open ODBC connection to the source database}

\item{attribute.connection}{a git-connection object to the attributes}

\item{walloon.connection}{a git-connection object to the Walloon source data}
}
\value{
the species group constraint information
}
\description{
Read the species for the raw datasource, save them to the results database
}

