% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_location.R
\name{read_location}
\alias{read_location}
\title{read the dataset of locations from the database}
\usage{
read_location(develop = TRUE)
}
\arguments{
\item{develop}{Logical value. Indicates the location ton the results database}
}
\description{
read the dataset of locations from the database
}
\examples{
location <- read_location()
head(location)
}

