% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_observation.R
\name{read_observation}
\alias{read_observation}
\title{Read the Flemish observations from a species}
\usage{
read_observation(species.id, first.winter, last.winter, species.covered,
  flemish.channel)
}
\arguments{
\item{species.id}{The id of the species}

\item{first.winter}{The oldest winter from which the observation are relevant. Observation prior to \code{first.winter} are ignored}

\item{last.winter}{The most recent winter from which the observation are relevant. Observation after \code{last.winter} are ignored}

\item{species.covered}{A vector with codes of relevant methodes}

\item{flemish.channel}{An open ODBC connection to the source database}
}
\value{
A \code{data.frame} with observations. \code{Complete= 1} indicates that the entire location was surveyed.
}
\description{
The selection uses several constraints. The user defined constraints are \code{first.winter} and \code{species.covered}. Following contraints are imposed at the same time as the user defined contraints:
\itemize{
\item The observation is validated
\item The observation is a midmonthly observation
\item The observation is not 'unobserved'
\item The location is 'active' and the observation date is within the 'active' period of the location
\item The observation is between october and march
}
}
\examples{
\dontrun{
observation <- read_observation(
  species.id = 4860,
  first.winter = 1992,
  species.covered = c("w", "weg", "wegm", "wem", "st")
)
head(observation)
}
}
