% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read_observation_wallonia.R
\name{read_observation_wallonia}
\alias{read_observation_wallonia}
\title{Read the Wallon observations from a species}
\usage{
read_observation_wallonia(species.id, first.winter)
}
\arguments{
\item{species.id}{The id of the species}

\item{first.winter}{The oldest winter from which the observation are relevant. Observation prior to \code{first.winter} are ignored}
}
\description{
All available imported data is used. The only constraint is
}

