% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_specieslist.R
\name{read_specieslist}
\alias{read_specieslist}
\title{Read the relevant species list}
\usage{
read_specieslist(result_channel, flemish_channel, walloon_repo,
  import_date)
}
\arguments{
\item{flemish_channel}{a DBI connection to the Flemish database}

\item{walloon_repo}{a git_repository object to the Walloon source data}

\item{import_date}{The last date of the import}
}
\description{
Read the relevant species list
}
