% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read_specieslist.R
\name{read_specieslist}
\alias{read_specieslist}
\title{Read the species list and their constraints}
\usage{
read_specieslist()
}
\description{
Read the species list and their constraints
}
\examples{
species.list <- read_specieslist()
head(species.list$species)
head(species.list$species.constraint)
}

