% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_specieslist.R
\name{read_specieslist}
\alias{read_specieslist}
\title{Read the species list and their constraints}
\usage{
read_specieslist(result.channel, flemish.channel, attribute.connection,
  limit = TRUE)
}
\arguments{
\item{result.channel}{An open RODBC connection to the results database}

\item{flemish.channel}{An open ODBC connection to the source database}

\item{attribute.connection}{a git-connection object to the attributes}

\item{limit}{Return only species with explicit constraints (default = TRUE). Otherwise return all species in the database}
}
\description{
Read the species list and their constraints
}
\examples{
result.channel <- n2khelper::connect_result()
flemish.channel <- connect_flemish_source(result.channel = result.channel)
attribute.connection <- connect_attribute(
  result.channel = result.channel,
  username = "Someone",
  password = "xxxx"
)
species.list <- read_specieslist(
  result.channel = result.channel,
  flemish.channel = flemish.channel,
  attribute.connection = attribute.connection
)
head(species.list$species)
head(species.list$species.constraint)
}

